refactor: move layout and page to components, create services directory for API calls
