+++ client/src/app/components/page.tsx
@@ -1,4 +1,4 @@
-import Header from "./components/layout/header";
+import Header from "./layout/header";
 
 export default function Home() {
   return (
