+++ client/src/app/newpost/page.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { FormEvent } from "react";
 import Header from "../components/layout/header";
 import PostForm from "./postform";
-import { getApiUrl } from '../utils/api';
+import { getApiUrl } from '../services/api';
 
 function NewPost() {
     const [title, setTitle] = useState("");
